obstacles
added cone obstacles that spawn and despawn with the road. they are triggers so that when we make the other cars they dont collide but they slow down the car and make the cars graphics flash to signify. the cones have no texture right now. the cones destroy when you hit them. they randomly spawn across the road.