implemented car AI
inside the traffic destroyer script the behavior is randomly decided, cars willl do 1 of 3 things: 0= yellow car that goes straight, 1 = red car that swerves away from you, 2=purple car that tries to hit you. i couldnt figure out a good way to assign the cars colors so on the traffic prefab i just added all the materials and use the indexes of them to assign color based on what behavior the car will have so they need to be on there, the only problem it seemed to cause was that the headlights arent red now for some reason but i couldnt find out a better way to do it right now.  theres also a variable inside the traffic destroyer script that you can change in start to make the cars turn less or more hard. theres also a limit to how close you have to be before the cars start moving.